Add repeat_with.rs file
